lawen-ejb/src/main/java/sessionbeans/TipoPatologiaFacade.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package sessionbeans;

import entities.TipoPatologia;
import javax.ejb.Stateless;
import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;

/**
 *
 * @author daniel
 */
@Stateless
public class TipoPatologiaFacade extends AbstractFacade<TipoPatologia> implements TipoPatologiaFacadeLocal {
    @PersistenceContext(unitName = "pingeso_lawen-ejb_ejb_1.0PU")
    private EntityManager em;

    @Override
     /**
     * Retorna el entity manager
     */
    protected EntityManager getEntityManager() {
        return em;
    }

    public void setEm(EntityManager em) {
        this.em = em;
    }
    
     /**
     * Super: obtener subclass
     */
    public TipoPatologiaFacade() {
        super(TipoPatologia.class);
    }
    
}
